Add density sort for each alphabetical character
The current implementation appears to be pretty inefficient due to the
fact that the getpixel() might be really inefficient based on the
documentation. I've since given up on trying to use the ImageMath module
to more efficiently count the number of black pixels in the glyph image.
It appears that the evaluation method runs in a completely separate
python environment and it doesn't seem possible to pass in a non-Image
variable into the method.

Until I learn more about PIL, or until such functionality is
implemented, this crude method is the only viable solution I have,
unless I frankenstein some stuff with another image library like OpenCV
(if I can manage the get that working in my Pycharm IDE). Future commits
will implement a way to properly crop the glyph onto the image canvas,
and keeping that standardized.